Refatoração da pasta shared Separação de responsabilidade das funções
