+++ shared/http/errorMessages.js
@@ -1,0 +1,13 @@
+module.exports = {
+    INVALID_TOKEN: "Chave de acesso inválida ou inexistente.",
+    INVALID_REQUEST: "Requisição inválida",
+    EMPTY_REQUEST: "É necessário o envio dos dados para realizar esta operação.",
+    INTERNAL_ERROR: "Erro interno ao realizar a operação.",
+    ID_REQUIRED: "O ID do registro é necessário para realizar esta consulta/atualização.",
+    DEFAULT_ERROR: "Ocorreu um erro durante a requisição.",
+    VALIDATION_ERROR_FOUND: "Erros de validação encontrados.",
+    USERNAME_EMAIL_ALREADY_USED: "Email e/ou nome de usário já cadastrado.",
+    USER_NOT_FOUND: "Usuário não encontrado",
+    TOKEN_ALREADY_GENERATED: "Um token já foi cadastrado para este usuário.",
+    INVALID_PASSWORD: "Senha incorreta."
+}
