+++ shared/http/infoMessages.js
@@ -1,0 +1,6 @@
+module.exports = {
+    SUCCESSFULLY_REGISTERED: "Usuário cadastrado com sucesso!.",
+    SUCCESSFULLY_REGISTERED_TOKEN: "Token de Acesso criado com sucesso!.",
+    SUCCESSFULLY_REQUEST: "Requisição realizada com sucesso!.",
+    SUCCESSFULLY_AUTHENTICATION: "Authenticado com sucesso!."
+}
