+++ shared/http/responseUtils.js
@@ -1,0 +1,9 @@
+ exports.createResponse = function (success, authorized, message, data) {
+    let response = {
+        success: success,
+        authorized: authorized,
+        message: message,
+        data: data
+    }
+    return response;
+}
