+++ shared/schemas/PlataformOverview.js
